+++ geo-types/CHANGES.md
@@ -21,8 +21,7 @@
   * Remove `set_lat()` -- use `set_y()` instead.
 * BREAKING: Remove deprecated `Polygon<T>::is_convex()` -- use `geo::is_convex` on `poly.exterior()` instead.
 * BREAKING: Remove deprecated `Rect<T>::try_new()` -- use `Rect::new` instead, since `Rect::try_new` will never Error. Also removes corresponding `InvalidRectCoordinatesError`.
-* BREAKING: Remove `GeometryCollection::new(value)`,  replace it with `GeometryCollection::new(value)`, and deprecate `GeometryCollection::new_from(value)`.
-  * The old `GeometryCollection::new()` is no longer available, and should be replaced with `GeometryCollection::default()` if an empty `GeometryCollection` is needed.
+* BREAKING: Replace deprecated `GeometryCollection::new()` with `GeometryCollection::new(value)`, and remove deprecated `GeometryCollection::new_from(value)`.
 
 ## 0.7.4